INITIAL

+++ kmod/core.c
@@ -79,9 +79,8 @@ static inline u64 BE64(u64 x) { return __builtin_bswap64(x); }
 
 //
 #define PORT(cid, pid) (XSOCK_SERVER_PORT + (cid)*10 + (pid))
-// WILL UNSIGNED OVERFLOW IF LOWER
-#define PORT_CID(port) (((uint)(port) - XSOCK_SERVER_PORT) / 10)
-#define PORT_PID(port) (((uint)(port) - XSOCK_SERVER_PORT) % 10)
+#define PORT_CID(port) (((port) - XSOCK_SERVER_PORT) / 10)
+#define PORT_PID(port) (((port) - XSOCK_SERVER_PORT) % 10)
 
 //
 #if XSOCK_SERVER && PORT(XSOCK_CONNS_N - 1, XSOCK_PATHS_N - 1) > 0xFFFF
